Move email_changed code into concern
